Refactor 4.2.0

+++ Otlob/Services/MealService.cs
@@ -2,16 +2,16 @@
 
 public class MealService(IMapper mapper, IFileService imageService,
                          IUnitOfWorkRepository unitOfWorkRepository,
-                         IRestaurantService restaurantService,
                          IEncryptionService encryptionService,
                          IMealPriceHistoryService mealPriceHistoryService, 
-                         IMealOptionGroupService optionGroupService) : IMealService
+                         IMealOptionGroupService optionGroupService,
+                         IManyMealsManyAddOnsService manyMealsManyAddOnsService) : IMealService
 {
     private readonly IMapper _mapper = mapper;
     private readonly IFileService _imageService = imageService;
-    private readonly IRestaurantService _restaurantService = restaurantService;
     private readonly IEncryptionService _encryptionService = encryptionService;
     private readonly IMealOptionGroupService _optionGroupService = optionGroupService;
+    private readonly IManyMealsManyAddOnsService _manyMealsManyAddOnsService = manyMealsManyAddOnsService;
     private readonly IUnitOfWorkRepository _unitOfWorkRepository = unitOfWorkRepository;
     private readonly IMealPriceHistoryService _mealPriceHistoryService = mealPriceHistoryService;
 
@@ -27,6 +27,7 @@ public class MealService(IMapper mapper, IFileService imageService,
                     Key = m.Id,
                     Name = m.Name,
                     Price = m.Price,
+                    OfferPrice = m.OfferPrice,
                     Image = m.Image,
                 }
              )!;
@@ -49,6 +50,7 @@ public class MealService(IMapper mapper, IFileService imageService,
                     Key = m.Id,
                     Name = m.Name,
                     Price = m.Price,
+                    OfferPrice = m.OfferPrice,
                     Image = m.Image,
                 }
              )!;
@@ -83,24 +85,27 @@ public class MealService(IMapper mapper, IFileService imageService,
         if (result.IsFailure)
             return Result.Failure(result.Error);
 
-        var isImageUploaded = _imageService.UploadImage(imageRequest.Image)!;
+        var imagePath = _imageService.UploadImage(imageRequest.Image)!;
 
-        if (isImageUploaded.IsFailure)
-            return Result.Failure(isImageUploaded.Error);
-
-        Meal meal = FillMealData(request, restaurantId, categoryId, isImageUploaded.Value);
-
+        Meal meal = _mapper.Map<Meal>((restaurantId, categoryId, request, imagePath.Value));
 
         // Add meal option groups and items if exist
-        if (request.HasOptionGroup && request.OptionGroups is not null)
+        if (request.HasOptionGroup && request.OptionGroups is not null && request.OptionGroups.Count > 0)
         {
-            var addOptionsResult = await _optionGroupService.Add(request.OptionGroups, meal.Id);
+            var addOptionsResult = await _optionGroupService.AddAsync(request.OptionGroups, meal.Id);
             
             if (addOptionsResult.IsFailure)
                 return Result.Failure(addOptionsResult.Error);
         }
 
         // add AddOns if exist
+        if (request.HasAddOns && request.SelectedAddOns is not null && request.SelectedAddOns.Count > 0)
+        {
+            var addAddOnsResult = await _manyMealsManyAddOnsService.AddAsync(request, restaurantId, meal.Id);
+
+            if (addAddOnsResult.IsFailure)
+                return Result.Failure(addAddOnsResult.Error);
+        }
 
         _mealPriceHistoryService.AddMealPriceHistory(meal.Id, meal.Price);
 
@@ -128,13 +133,22 @@ public class MealService(IMapper mapper, IFileService imageService,
             (request.HasOptionGroup && !oldMeal.HasOptionGroup) ||
             (request.HasOptionGroup && oldMeal.HasOptionGroup))
         {
-            var updateOptionsResult = await _optionGroupService.Update(request.OptionGroups!, mealId);
+            var updateOptionsResult = await _optionGroupService.UpdateAsync(request.OptionGroups!, mealId);
             
             if (updateOptionsResult.IsFailure)
                 return Result.Failure(updateOptionsResult.Error);
         }
 
         // update AddOns if exist
+        if ((!request.HasAddOns && oldMeal.HasAddOn) ||
+            (request.HasAddOns && !oldMeal.HasAddOn) ||
+            (request.HasAddOns && oldMeal.HasAddOn))
+        { 
+            var updateAddOnsResult = await _manyMealsManyAddOnsService.UpdateAsync(request, restaurantId, mealId);
+
+            if (updateAddOnsResult.IsFailure)
+                return Result.Failure(updateAddOnsResult.Error);
+        }
 
         if (oldMeal.Price != request.Price)
             _mealPriceHistoryService.UpdateMealPriceHistory(mealId, request.Price);
@@ -161,6 +175,9 @@ public class MealService(IMapper mapper, IFileService imageService,
 
         Meal meal = GetMealImageById(mealId);
 
+        if (meal.Image is null)
+            return Result.Failure(MealErrors.MealNotFound);
+
         _imageService.DeleteImage(meal.Image);
 
         meal.Image = uploadImageResult.Value;
@@ -170,7 +187,7 @@ public class MealService(IMapper mapper, IFileService imageService,
         _unitOfWorkRepository.SaveChanges();
 
         return Result.Success();
-    }    
+    }
 
     public Result<IQueryable<MealResponse>> GetDeletedMeals(int restaurantId)
     {
@@ -186,6 +203,7 @@ public class MealService(IMapper mapper, IFileService imageService,
                     Name = m.Name,
                     Image = m.Image,
                     Price = m.Price,
+                    OfferPrice = m.OfferPrice
                 }
              )!;
 
@@ -195,7 +213,8 @@ public class MealService(IMapper mapper, IFileService imageService,
 
         return Result.Success(response);
     }
-
+    
+    // TODO: Make Delete in one method toggle delete/unDelete
     public Result DeleteMeal(string mealId)
     {
         Result validateMealIdResult = IsMealIdExists(mealId);
@@ -216,6 +235,20 @@ public class MealService(IMapper mapper, IFileService imageService,
         {
             _unitOfWorkRepository.MealOptionItems.SoftDelete(expression: oi => mealOptionGroupIds.Contains(oi.OptionGroupId));
             _unitOfWorkRepository.MealOptionGroups.SoftDelete(expression: og => og.MealId == mealId);
+        }
+
+        var mealAddOnsIds = _unitOfWorkRepository.ManyMealManyAddOns
+            .GetAllWithSelect
+             (
+                expression: mao => mao.MealId == mealId,
+                tracked: false,
+                selector: mao => mao.AddOnId
+             )!
+             .ToList();
+
+        if (mealAddOnsIds.Count > 0)
+        {
+            _unitOfWorkRepository.ManyMealManyAddOns.SoftDelete(expression: mao => mao.MealId == mealId);
         }
 
         _unitOfWorkRepository.Meals.SoftDelete(expression: m => m.Id == mealId);
@@ -248,10 +281,125 @@ public class MealService(IMapper mapper, IFileService imageService,
             _unitOfWorkRepository.MealOptionGroups.UnSoftDelete(expression: og => og.MealId == mealId);
         }
 
+        var mealAddOnsIds = _unitOfWorkRepository.ManyMealManyAddOns
+            .GetAllWithSelect
+             (
+                expression: mao => mao.MealId == mealId,
+                tracked: false,
+                ignoreQueryFilter: true,
+                selector: mao => mao.AddOnId
+             )!
+             .ToList();
+
+        if (mealAddOnsIds.Count > 0)
+        {
+            _unitOfWorkRepository.ManyMealManyAddOns.UnSoftDelete(expression: mao => mao.MealId == mealId);
+        }
+
         _unitOfWorkRepository.Meals.UnSoftDelete(expression: m => m.Id == mealId);
         
         _unitOfWorkRepository.SaveChanges();
         
+        return Result.Success();
+    }
+
+    // TODO: Make Delete in one method toggle delete/unDelete
+    public Result DeleteManyMealsByCategoryId(int categoryId)
+    {
+        var mealsIds = _unitOfWorkRepository.Meals
+            .GetAllWithSelect
+             (
+                expression: m => m.CategoryId == categoryId,
+                tracked: false,
+                selector: m => m.Id
+             )!
+             .ToList();
+
+        if (mealsIds.Count > 0)
+        {
+            var mealOptionGroupIds = _unitOfWorkRepository.MealOptionGroups
+                .GetAllWithSelect
+                 (
+                    expression: og => mealsIds.Contains(og.MealId),
+                    tracked: false,
+                    selector: og => og.MealOptionGroupId
+                 )!
+                 .ToList();
+
+            if (mealOptionGroupIds.Count > 0)
+            {
+                _unitOfWorkRepository.MealOptionItems.SoftDelete(expression: oi => mealOptionGroupIds.Contains(oi.OptionGroupId));
+                _unitOfWorkRepository.MealOptionGroups.SoftDelete(expression: og => mealsIds.Contains(og.MealId));
+            }
+
+            var mealAddOnsIds = _unitOfWorkRepository.ManyMealManyAddOns
+                .GetAllWithSelect
+                 (
+                    expression: mao => mealsIds.Contains(mao.MealId),
+                    tracked: false,
+                    selector: mao => mao.AddOnId
+                 )!
+                 .ToList();
+
+            if (mealAddOnsIds.Count > 0)
+            {
+                _unitOfWorkRepository.ManyMealManyAddOns.SoftDelete(expression: mao => mealsIds.Contains(mao.MealId));
+            }
+
+            _unitOfWorkRepository.Meals.SoftDelete(expression: m => m.CategoryId == categoryId);
+        }
+
+        return Result.Success();
+    }
+
+    public Result UnDeleteManyMealsByCategoryId(int categoryId)
+    {
+        var mealsIds = _unitOfWorkRepository.Meals
+            .GetAllWithSelect
+             (
+                expression: m => m.CategoryId == categoryId,
+                tracked: false,
+                ignoreQueryFilter: true,
+                selector: m => m.Id
+             )!
+             .ToList();
+
+        if (mealsIds.Count > 0)
+        {
+            var mealOptionGroupIds = _unitOfWorkRepository.MealOptionGroups
+                .GetAllWithSelect
+                 (
+                    expression: og => mealsIds.Contains(og.MealId),
+                    tracked: false,
+                    ignoreQueryFilter: true,
+                    selector: og => og.MealOptionGroupId
+                 )!
+                 .ToList();
+
+            if (mealOptionGroupIds.Count > 0)
+            {
+                _unitOfWorkRepository.MealOptionItems.UnSoftDelete(expression: oi => mealOptionGroupIds.Contains(oi.OptionGroupId));            
+                _unitOfWorkRepository.MealOptionGroups.UnSoftDelete(expression: og => mealsIds.Contains(og.MealId));
+            }
+
+            var mealAddOnsIds = _unitOfWorkRepository.ManyMealManyAddOns
+                .GetAllWithSelect
+                 (
+                    expression: mao => mealsIds.Contains(mao.MealId),
+                    tracked: false,
+                    ignoreQueryFilter: true,
+                    selector: mao => mao.AddOnId
+                 )!
+                 .ToList();
+
+            if (mealAddOnsIds.Count > 0)
+            {
+                _unitOfWorkRepository.ManyMealManyAddOns.UnSoftDelete(expression: mao => mealsIds.Contains(mao.MealId));
+            }
+
+            _unitOfWorkRepository.Meals.UnSoftDelete(expression: m => m.CategoryId == categoryId);
+        }
+
         return Result.Success();
     }
 
@@ -275,9 +423,9 @@ public class MealService(IMapper mapper, IFileService imageService,
     
     private Result ValidateMealOnAdd(int restaurantId, int categoryId, MealRequest request)
     {
-        Result isRestaurantIdExists = _restaurantService.IsRestaurantIdExists(restaurantId);
+        bool isRestaurantIdExists = _unitOfWorkRepository.Restaurants.IsExist(r => r.Id == restaurantId);
 
-        if (isRestaurantIdExists.IsFailure)
+        if (!isRestaurantIdExists)
             return Result.Failure(RestaurantErrors.NotFound);
 
         bool isCategoryIdExists = _unitOfWorkRepository.MealCategories.IsExist(mc => mc.Id == categoryId);
@@ -333,20 +481,10 @@ public class MealService(IMapper mapper, IFileService imageService,
                 oldMeal.CategoryId == categoryId && oldMeal.NumberOfServings == request.NumberOfServings;
     }
 
-    private Meal FillMealData(MealRequest request, int restaurantId, int categoryId, string imageUrl)
-    {
-        Meal meal = _mapper.Map<Meal>(request);
-
-        meal.RestaurantId = restaurantId;
-        meal.CategoryId = categoryId;
-        meal.Image = imageUrl;
-
-        return meal;
-    }
-
     private MealResponse FillResponseData(Meal meal)
     {
         var response = _mapper.Map<MealResponse>(meal);
+        
         response.CategoryKey = _encryptionService.Encrypt(meal!.CategoryId);
 
         response.Categories = _unitOfWorkRepository.MealCategories
@@ -361,37 +499,65 @@ public class MealService(IMapper mapper, IFileService imageService,
                 }
              );
 
-        response.OptionGroups = _unitOfWorkRepository.MealOptionGroups
+        if (meal.HasOptionGroup)
+        {
+            response.OptionGroups = _unitOfWorkRepository.MealOptionGroups
+                .GetAllWithSelect
+                 (
+                    expression: og => og.MealId == meal.Id,
+                    tracked: false,
+                    selector: og => new OptionGroupResponse
+                    {
+                        Id = og.MealOptionGroupId,
+                        Name = og.Name,
+                        DisplayOrder = og.DisplayOrder,
+                        OptionItems = og.OptionItems.Select(oi => new OptionItemResponse
+                        {
+                            Id = oi.MealOptionItemId,
+                            Name = oi.Name,
+                            Price = oi.Price,
+                            Image = oi.Image,
+                            DisplayOrder = oi.DisplayOrder,
+                            IsPobular = oi.IsPobular
+                        })
+                    }
+                 )!;
+
+            if (response.OptionGroups.Any())
+            {
+                response.OptionGroups = response.OptionGroups
+                    .Select(og =>
+                    {
+                        og.OptionItems = og.OptionItems.OrderBy(oi => oi.DisplayOrder);
+                        return og;
+                    })
+                    .OrderBy(og => og.DisplayOrder);
+            }
+        }
+
+        response.AddOns = _unitOfWorkRepository.MealAddOns
             .GetAllWithSelect
              (
-                expression: og => og.MealId == meal.Id,
+                expression: ao => ao.RestaurantId == meal.RestaurantId,
                 tracked: false,
-                selector: og => new OptionGroupResponse
+                selector: ao => new AddOnResponse
                 {
-                    Id = og.MealOptionGroupId,
-                    Name = og.Name,
-                    DisplayOrder = og.DisplayOrder,
-                    OptionItems = og.OptionItems.Select(oi => new OptionItemResponse
-                    {
-                        Id = oi.MealOptionItemId,
-                        Name = oi.Name,
-                        Price = oi.Price,
-                        Image = oi.Image,
-                        DisplayOrder = oi.DisplayOrder,
-                        IsPobular = oi.IsPobular
-                    })                    
+                    Key = ao.Id,
+                    Name = ao.Name,
+                    Price = ao.Price,
+                    Image = ao.Image
                 }
              )!;
 
-        if (response.OptionGroups.Any())
+        if (meal.HasAddOn)
         {
-            response.OptionGroups = response.OptionGroups
-                .Select(og =>
-                {
-                    og.OptionItems = og.OptionItems.OrderBy(oi => oi.DisplayOrder);
-                    return og;
-                })
-                .OrderBy(og => og.DisplayOrder);
+            response.SelectedAddOns = _unitOfWorkRepository.ManyMealManyAddOns
+                .GetAllWithSelect
+                 (
+                    expression: mao => mao.MealId == meal.Id,
+                    tracked: false,
+                    selector: mao => mao.AddOnId
+                 )!;
         }
 
         return response;
